[separate_mailchimp_log] Renamed classes to make clear functionality
